Combine insect hooks
