+++ src/Modules/Insects/Butterfly.cs
@@ -1,100 +1,7 @@
 ﻿using Random = UnityEngine.Random;
-using DevInterface;
 
 namespace RegionKit.Modules.Insects
 {
-	internal static class ButterfliesCI
-	{
-		internal static void Apply()
-		{
-			On.InsectCoordinator.CreateInsect += InsectCoordinator_CreateInsect;
-			On.InsectCoordinator.TileLegalForInsect += InsectCoordinator_TileLegalForInsect;
-			On.InsectCoordinator.EffectSpawnChanceForInsect += InsectCoordinator_EffectSpawnChanceForInsect;
-			On.InsectCoordinator.RoomEffectToInsectType += InsectCoordinator_RoomEffectToInsectType;
-			_CommonHooks.PostRoomLoad += PostRoomLoad;
-		}
-
-		internal static void Undo()
-		{
-			On.InsectCoordinator.CreateInsect -= InsectCoordinator_CreateInsect;
-			On.InsectCoordinator.TileLegalForInsect -= InsectCoordinator_TileLegalForInsect;
-			On.InsectCoordinator.EffectSpawnChanceForInsect -= InsectCoordinator_EffectSpawnChanceForInsect;
-			On.InsectCoordinator.RoomEffectToInsectType -= InsectCoordinator_RoomEffectToInsectType;
-			_CommonHooks.PostRoomLoad -= PostRoomLoad;
-		}
-
-
-		private static bool InsectCoordinator_TileLegalForInsect(On.InsectCoordinator.orig_TileLegalForInsect orig, CosmeticInsect.Type type, Room room, Vector2 testPos)
-		{
-			bool aiNoNarrow = !room.readyForAI || !room.aimap.getAItile(testPos).narrowSpace;
-			return (type == _Enums.ButterflyA || type == _Enums.ButterflyB) ? !room.GetTile(testPos).AnyWater && aiNoNarrow : orig(type, room, testPos);
-		}
-
-		private static void PostRoomLoad(Room self)
-		{
-			for (int i = 0; i < self.roomSettings.effects.Count; i++)
-			{
-				if (self.roomSettings.effects[i].type == _Enums.ButterfliesA || self.roomSettings.effects[i].type == _Enums.ButterfliesB)
-				{
-					if (self.insectCoordinator == null)
-					{
-						self.insectCoordinator = new InsectCoordinator(self);
-						self.AddObject(self.insectCoordinator);
-					}
-					self.insectCoordinator.AddEffect(self.roomSettings.effects[i]);
-				}
-			}
-		}
-
-		private static CosmeticInsect.Type InsectCoordinator_RoomEffectToInsectType(On.InsectCoordinator.orig_RoomEffectToInsectType orig, RoomSettings.RoomEffect.Type type)
-		{
-			if (type == _Enums.ButterfliesA)
-				return _Enums.ButterflyA;
-			else if (type == _Enums.ButterfliesB)
-				return _Enums.ButterflyB;
-			else
-				return orig(type);
-		}
-
-		private static void InsectCoordinator_CreateInsect(On.InsectCoordinator.orig_CreateInsect orig, InsectCoordinator self, CosmeticInsect.Type type, Vector2 pos, InsectCoordinator.Swarm swarm)
-		{
-			if (!InsectCoordinator.TileLegalForInsect(type, self.room, pos))
-			{
-				return;
-			}
-			if (self.room.world.rainCycle.TimeUntilRain < Random.Range(1200, 1600))
-			{
-				return;
-			}
-
-			if (type == _Enums.ButterflyA || type == _Enums.ButterflyB)
-			{
-				CosmeticInsect insect = new Butterfly(self.room, pos, type == _Enums.ButterflyA);
-
-				self.allInsects.Add(insect);
-				if (swarm != null)
-				{
-					swarm.members.Add(insect);
-					insect.mySwarm = swarm;
-				}
-				self.room.AddObject(insect);
-			}
-			else
-			{
-				orig(self, type, pos, swarm);
-			}
-		}
-
-		private static bool InsectCoordinator_EffectSpawnChanceForInsect(On.InsectCoordinator.orig_EffectSpawnChanceForInsect orig, CosmeticInsect.Type type, Room room, Vector2 testPos, float effectAmount)
-		{
-			if (type == _Enums.ButterflyA || type == _Enums.ButterflyB)
-			{
-				return Mathf.Pow(Random.value, 1f - effectAmount) > (room.readyForAI ? room.aimap.getTerrainProximity(testPos) : 5) * 0.05f;
-			}
-			return orig(type, room, testPos, effectAmount);
-		}
-	}
-
 	/// <summary>
 	/// By Alduris
 	/// Moth-like creature.
