+++ src/Modules/Insects/GlowingSwimmer.cs
@@ -4,92 +4,6 @@ namespace RegionKit.Modules.Insects;
 
 /// <summary>
 /// By LB/M4rbleL1ne
-/// Glowing swimmer insect
-/// </summary>
-internal static class GlowingSwimmersCI
-{
-	internal static void Apply()
-	{
-		_CommonHooks.PostRoomLoad += PostRoomLoad;
-		On.InsectCoordinator.SpeciesDensity_Type_1 += InsectCoordinatorSpeciesDensityInsectType;
-		On.InsectCoordinator.RoomEffectToInsectType += InsectCoordinatorRoomEffectToInsectType;
-		On.InsectCoordinator.TileLegalForInsect += InsectCoordinatorTileLegalForInsect;
-		On.InsectCoordinator.EffectSpawnChanceForInsect += InsectCoordinatorEffectSpawnChanceForInsect;
-		On.InsectCoordinator.CreateInsect += InsectCoordinatorCreateInsect;
-	}
-
-	internal static void Undo()
-	{
-		_CommonHooks.PostRoomLoad -= PostRoomLoad;
-		On.InsectCoordinator.SpeciesDensity_Type_1 -= InsectCoordinatorSpeciesDensityInsectType;
-		On.InsectCoordinator.RoomEffectToInsectType -= InsectCoordinatorRoomEffectToInsectType;
-		On.InsectCoordinator.TileLegalForInsect -= InsectCoordinatorTileLegalForInsect;
-		On.InsectCoordinator.EffectSpawnChanceForInsect -= InsectCoordinatorEffectSpawnChanceForInsect;
-		On.InsectCoordinator.CreateInsect -= InsectCoordinatorCreateInsect;
-	}
-
-	private static void PostRoomLoad(Room self)
-	{
-		List<RoomSettings.RoomEffect> efs = self.roomSettings.effects;
-		for (var i = 0; i < efs.Count; i++)
-		{
-			RoomSettings.RoomEffect effect = efs[i];
-			if (effect.type == _Enums.GlowingSwimmers)
-			{
-				if (self.insectCoordinator is null)
-				{
-					self.insectCoordinator = new(self);
-					self.AddObject(self.insectCoordinator);
-				}
-				self.insectCoordinator.AddEffect(effect);
-				break;
-			}
-		}
-	}
-
-	private static float InsectCoordinatorSpeciesDensityInsectType(On.InsectCoordinator.orig_SpeciesDensity_Type_1 orig, CosmeticInsect.Type type)
-	{
-		return type == _Enums.GlowingSwimmerInsect ? .8f : orig(type);
-	}
-
-	private static CosmeticInsect.Type InsectCoordinatorRoomEffectToInsectType(On.InsectCoordinator.orig_RoomEffectToInsectType orig, RoomSettings.RoomEffect.Type type)
-	{
-		return type == _Enums.GlowingSwimmers ? _Enums.GlowingSwimmerInsect : orig(type);
-	}
-
-	private static bool InsectCoordinatorTileLegalForInsect(On.InsectCoordinator.orig_TileLegalForInsect orig, CosmeticInsect.Type type, Room room, Vector2 testPos)
-	{
-		return type == _Enums.GlowingSwimmerInsect ? room.GetTile(testPos).DeepWater : orig(type, room, testPos);
-	}
-
-	private static bool InsectCoordinatorEffectSpawnChanceForInsect(On.InsectCoordinator.orig_EffectSpawnChanceForInsect orig, CosmeticInsect.Type type, Room room, Vector2 testPos, float effectAmount)
-	{
-		return type == _Enums.GlowingSwimmerInsect || orig(type, room, testPos, effectAmount);
-	}
-
-	private static void InsectCoordinatorCreateInsect(On.InsectCoordinator.orig_CreateInsect orig, InsectCoordinator self, CosmeticInsect.Type type, Vector2 pos, InsectCoordinator.Swarm swarm)
-	{
-		if (type == _Enums.GlowingSwimmerInsect)
-		{
-			Room rm = self.room;
-			if (!InsectCoordinator.TileLegalForInsect(type, rm, pos) || rm.world.rainCycle.TimeUntilRain < UnityEngine.Random.Range(1200, 1600))
-				return;
-			var cosmeticInsect = new GlowingSwimmer(rm, pos);
-			self.allInsects.Add(cosmeticInsect);
-			if (swarm is not null)
-			{
-				swarm.members.Add(cosmeticInsect);
-				cosmeticInsect.mySwarm = swarm;
-			}
-			rm.AddObject(cosmeticInsect);
-		}
-		else
-			orig(self, type, pos, swarm);
-	}
-}
-
-/// <summary>
-/// By LB/M4rbleL1ne
 /// Glowing swimmer
 /// </summary>
 public class GlowingSwimmer : CosmeticInsect
