+++ src/Modules/Insects/Seedling.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RegionKit.Modules.Insects
+{
+	public class Seedling : CosmeticInsect
+	{
+		public Seedling(Room room, Vector2 pos, Type type) : base(room, pos, type)
+		{
+		}
+	}
+}
