+++ src/Modules/Insects/Zipper.cs
@@ -2,92 +2,6 @@
 
 namespace RegionKit.Modules.Insects
 {
-	internal static class ZippersCI
-	{
-		internal static void Apply()
-		{
-			On.InsectCoordinator.CreateInsect += InsectCoordinator_CreateInsect;
-			On.InsectCoordinator.TileLegalForInsect += InsectCoordinator_TileLegalForInsect;
-			On.InsectCoordinator.EffectSpawnChanceForInsect += InsectCoordinator_EffectSpawnChanceForInsect;
-			On.InsectCoordinator.RoomEffectToInsectType += InsectCoordinator_RoomEffectToInsectType;
-			_CommonHooks.PostRoomLoad += PostRoomLoad;
-		}
-
-		internal static void Undo()
-		{
-			On.InsectCoordinator.CreateInsect -= InsectCoordinator_CreateInsect;
-			On.InsectCoordinator.TileLegalForInsect -= InsectCoordinator_TileLegalForInsect;
-			On.InsectCoordinator.EffectSpawnChanceForInsect -= InsectCoordinator_EffectSpawnChanceForInsect;
-			On.InsectCoordinator.RoomEffectToInsectType -= InsectCoordinator_RoomEffectToInsectType;
-			_CommonHooks.PostRoomLoad -= PostRoomLoad;
-		}
-
-
-		private static bool InsectCoordinator_TileLegalForInsect(On.InsectCoordinator.orig_TileLegalForInsect orig, CosmeticInsect.Type type, Room room, Vector2 testPos)
-		{
-			return (type == _Enums.Zipper) ? !room.GetTile(testPos).AnyWater : orig(type, room, testPos);
-		}
-
-		private static void PostRoomLoad(Room self)
-		{
-			for (int i = 0; i < self.roomSettings.effects.Count; i++)
-			{
-				if (self.roomSettings.effects[i].type == _Enums.Zippers)
-				{
-					if (self.insectCoordinator == null)
-					{
-						self.insectCoordinator = new InsectCoordinator(self);
-						self.AddObject(self.insectCoordinator);
-					}
-					self.insectCoordinator.AddEffect(self.roomSettings.effects[i]);
-				}
-			}
-		}
-
-		private static CosmeticInsect.Type InsectCoordinator_RoomEffectToInsectType(On.InsectCoordinator.orig_RoomEffectToInsectType orig, RoomSettings.RoomEffect.Type type)
-		{
-			return type == _Enums.Zippers ? _Enums.Zipper : orig(type);
-		}
-
-		private static void InsectCoordinator_CreateInsect(On.InsectCoordinator.orig_CreateInsect orig, InsectCoordinator self, CosmeticInsect.Type type, Vector2 pos, InsectCoordinator.Swarm swarm)
-		{
-			if (!InsectCoordinator.TileLegalForInsect(type, self.room, pos))
-			{
-				return;
-			}
-			if (self.room.world.rainCycle.TimeUntilRain < Random.Range(1200, 1600))
-			{
-				return;
-			}
-
-			if (type == _Enums.Zipper)
-			{
-				CosmeticInsect insect = new Zipper(self.room, pos);
-
-				self.allInsects.Add(insect);
-				if (swarm != null)
-				{
-					swarm.members.Add(insect);
-					insect.mySwarm = swarm;
-				}
-				self.room.AddObject(insect);
-			}
-			else
-			{
-				orig(self, type, pos, swarm);
-			}
-		}
-
-		private static bool InsectCoordinator_EffectSpawnChanceForInsect(On.InsectCoordinator.orig_EffectSpawnChanceForInsect orig, CosmeticInsect.Type type, Room room, Vector2 testPos, float effectAmount)
-		{
-			if (type == _Enums.Zipper)
-			{
-				return Mathf.Pow(Random.value, 1f - effectAmount) > (room.readyForAI ? room.aimap.getTerrainProximity(testPos) : 5) * 0.05f;
-			}
-			return orig(type, room, testPos, effectAmount);
-		}
-	}
-
 	/// <summary>
 	/// By Alduris
 	/// Fly-like that zips around brightly sometimes
