+++ src/Modules/Insects/_Enums.cs
@@ -26,6 +26,10 @@
 		/// Zipper effect enum
 		/// </summary>
 		public static RoomSettings.RoomEffect.Type Zippers = new(nameof(Zippers), true);
+		/// <summary>
+		/// Seedling effect enum
+		/// </summary>
+		public static RoomSettings.RoomEffect.Type Seedlings = new(nameof(Seedlings), true);
 
 		/// <summary>
 		/// Glowing Swimmer insect enum
@@ -51,10 +55,47 @@
 		/// Zipper enum
 		/// </summary>
 		public static CosmeticInsect.Type Zipper = new(nameof(Zipper), true);
+		/// <summary>
+		/// Seedling enum
+		/// </summary>
+		public static CosmeticInsect.Type Seedling = new(nameof(Seedling), true);
 
 		/// <summary>
 		/// Insect category
 		/// </summary>
 		public static DevInterface.RoomSettingsPage.DevEffectsCategories RegionKitInsects = new(nameof(RegionKitInsects), true);
+
+		/// <summary>
+		/// Returns whether or not the cosmetic insect type is from RegionKit
+		/// </summary>
+		/// <param name="type">The cosmetic insect type</param>
+		/// <returns>If the enum is from RegionKit</returns>
+
+		public static bool IsRegionKitInsect(this CosmeticInsect.Type type)
+		{
+			return type == GlowingSwimmerInsect
+				|| type == ColoredCamoBeetle
+				|| type == MosquitoInsect
+				|| type == ButterflyA
+				|| type == ButterflyB
+				|| type == Zipper
+				|| type == Seedling;
+		}
+
+		/// <summary>
+		/// Returns whether or not the room effect type is a cosmetic insect from RegionKit
+		/// </summary>
+		/// <param name="type">The room effect type</param>
+		/// <returns>If the enum is from RegionKit</returns>
+		public static bool IsRegionKitInsect(this RoomSettings.RoomEffect.Type type)
+		{
+			return type == GlowingSwimmers
+				|| type == ColoredCamoBeetles
+				|| type == MosquitoInsects
+				|| type == ButterfliesA
+				|| type == ButterfliesB
+				|| type == Zippers
+				|| type == Seedlings;
+		}
 	}
 }
